Jwt implementation

+++ bank-api/src/main/java/com/paul/bankapi/Account/Service/AccountServiceImpl.java
@@ -16,7 +16,7 @@ public class AccountServiceImpl implements AccountService{
     private final CustomerService customerService;
 
     @Override
-    public Account addAccount(Account account, Long customerId) {
+    public Account addAccount(Account account, Integer customerId) {
         Optional<Customer> optionalCustomer = customerService.findCustomerById(customerId);
         if (optionalCustomer.isPresent()) {
             Customer customer = optionalCustomer.get();
@@ -30,7 +30,7 @@ public class AccountServiceImpl implements AccountService{
     }
 
     @Override
-    public boolean removeAccount(Long accountId,Long customerId) {
+    public boolean removeAccount(Integer accountId,Integer customerId) {
         Optional<Customer> optionalCustomer = customerService.findCustomerById(customerId);
         if (optionalCustomer.isPresent()){
             Optional<Account> optionalAccount = accountRepository.findById(accountId);
@@ -44,7 +44,7 @@ public class AccountServiceImpl implements AccountService{
     }
 
     @Override
-    public Optional<Account> findAccountById(Long accountId) {
+    public Optional<Account> findAccountById(Integer accountId) {
         Optional<Account> optionalAccount = accountRepository.findById(accountId);
         if (optionalAccount.isPresent()) {
             Account account = optionalAccount.get();
